Added Limit()
added limit(). Can then also catch the last geohash viewed to load the next "20" / list. You can also store that in your firestore for later retrieval if you like or simply on the fly. catching geohash not implamented as how you want to do taht cn vary depending on yoru needs. but is so simple to do that there is no point in adding that.